another

+++ Palindrome.py
@@ -4,7 +4,7 @@ temp = num
 while (temp>0):
     rem = temp%10
     rev = rem + (rev*10)
-    temp = int(temp/10)
+    temp = temp/10
 if (rev==num):
     print("It is a Palindrome Number")
 else:
